Refactor public connection payload into canonical info struct
